pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caringcaribou"
version = "0.7"
description = "A friendly automotive security exploration tool"
readme = "README.md"
authors = [{ name = "Kasper Karlsson" }]
license = { file = "LICENSE" }
keywords = ["automotive", "security", "CAN", "automotive protocols", "fuzzing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
]
dependencies = ["python-can"]

[project.urls]
Homepage = "https://github.com/CaringCaribou/caringcaribou"
Download = "https://github.com/CaringCaribou/caringcaribou/archive/master.tar.gz"

[project.optional-dependencies]
test = ["pytest", "tox"]

[project.scripts]
caringcaribou = "caringcaribou.caringcaribou:main"

[project.entry-points."caringcaribou.modules"]
dcm = "caringcaribou.modules.dcm"
doip = "caringcaribou.modules.doip"
dump = "caringcaribou.modules.dump"
fuzzer = "caringcaribou.modules.fuzzer"
listener = "caringcaribou.modules.listener"
send = "caringcaribou.modules.send"
test = "caringcaribou.modules.test"
uds_fuzz = "caringcaribou.modules.uds_fuzz"
uds = "caringcaribou.modules.uds"
xcp = "caringcaribou.modules.xcp"

[tool.ruff]
# Set the maximum line length to 120.
line-length = 120